Rename product.tag to productTag
