--- server/src/validation-dtos/product.tag/product.tag.ts
@@ -1,1 +1,0 @@
-export{ProductTagDto} from "./product.tag.dto"
