+++ server/src/validation-dtos/productTag/productTag.ts
@@ -1,0 +1,1 @@
+export{ProductTagDto} from "./productTag.dto"
