+++ server/src/validation-dtos/validation-dtos.ts
@@ -1,4 +1,4 @@
 export {UserSignInValidationDto, UserSignUpValidationDto} from './user/user'
 export {ProductDto} from "./product/product"
 export {RecipeDto} from "./recipe/recipe"
-export {ProductTagDto} from "./product.tag/product.tag"
+export {ProductTagDto} from "./productTag/productTag"
